backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
rembg==2.0.54
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
python-dotenv==1.0.0
aiofiles==23.2.1
